+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `u`/`v` coordinates (offset from DCS map origin in meters) to `Position`s
 - `GetRealTime` API
 - Added `orientation` and `velocity` to `Weapon` object
+- Added DCS `time` of the update to units stream (`StreamUnitsResponse`)
 
 ### Changed
 - Unit objects now return the full group object in the `group` field to make event processing easier. This replaces the `group_name` and `group_category` fields and is a backwards incompatible change.